feat: fulfill Generation and Uid for Service

+++ internal/handler/db/sqlite/schema.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS services (
     name TEXT NOT NULL,
     description TEXT NOT NULL,
     uid TEXT NOT NULL,
-    generation INTEGER NOT NULL,
+    generation BIGINT NOT NULL,
     uri TEXT NOT NULL,
     created_at TIMESTAMP NOT NULL,
     PRIMARY KEY (parent, name)
